// apis/cars/types.ts
export interface CarType {
  car_length: string;
  color: string;
  drivetrain: string;
  engine: string;
  fuel_tank_capacity: string;
  fuel_type: string;
  height: string;
  kilometer: string;
  location: string;
  make: string;
  max_power: string;
  max_torque: string;
  model: string;
  owner: string;
  price: string;
  seating_capacity: string;
  seller_type: string;
  transmission: string;
  width: string;
  year: string;
}